mv `jointcaller` pkg to top-level
joint caller itself stays in `commands` with the others, for now!
