+++ src/test/scala/org/hammerlab/guacamole/main/SomaticJointCallerIntegrationTests.scala
@@ -1,9 +1,8 @@
 package org.hammerlab.guacamole.main
 
 import org.apache.spark.SparkContext
-import org.hammerlab.guacamole.commands.SparkCommand
-import org.hammerlab.guacamole.commands.jointcaller.SomaticJoint
-import org.hammerlab.guacamole.commands.jointcaller.SomaticJoint.Arguments
+import org.hammerlab.guacamole.commands.SomaticJoint.Arguments
+import org.hammerlab.guacamole.commands.{SomaticJoint, SparkCommand}
 import org.hammerlab.guacamole.data.{CancerWGSTestUtil, NA12878TestUtil}
 import org.hammerlab.guacamole.loci.set.LociSet
 import org.hammerlab.guacamole.util.TestUtil
